fix(folder): Prevent misleading warnings
Given where these warnings have been implemented, we are getting them in many recipes where light paths and receiver files have no meaning. And these lead to confusion among users and other developers. For now, I am just removing the warnings but, in the future, maybe we add them back in a place that is specific to the relevant recipes.